requirements.txt
asgiref==3.7.2
attr==0.3.2
Django==5.0.3
django-debug-toolbar==4.3.0
django-filter==24.1
djangorestframework==3.14.0
intervaltree==3.1.0
networkx==3.2.1
pillow==10.2.0
pytz==2024.1
sortedcontainers==2.4.0
sqlparse==0.4.4
tag==0.5
typing_extensions==4.10.0
